add helpers for file progress
